Adjust Onboarding to use Page Section
