+++ src/__tests__/src/ClimatePersonality.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
 
-import ClimatePersonality from '../../pages/ClimatePersonality';
+import StartQuiz from '../../pages/StartQuiz';
 
 // Mock react router to simulate history.push on button click
 jest.mock('react-router-dom', () => ({
@@ -13,7 +13,7 @@ window.scrollTo = jest.fn();
 
 describe('ClimatePersonality', () => {
   it('shows the main heading ', () => {
-    const { getByText } = render(<ClimatePersonality />);
+    const { getByText } = render(<StartQuiz />);
     expect(
       getByText(
         /We want to make constructive conversations about climate change easier./i
@@ -21,19 +21,19 @@ describe('ClimatePersonality', () => {
     ).toBeInTheDocument();
   });
   it('shows header core values heading ', () => {
-    const { getByText } = render(<ClimatePersonality />);
+    const { getByText } = render(<StartQuiz />);
     expect(getByText(/Let's find out your core values/i)).toBeInTheDocument();
   });
 
   it('It explains about the core values ', () => {
-    const { getByText } = render(<ClimatePersonality />);
+    const { getByText } = render(<StartQuiz />);
     expect(
       getByText(/By answering 10 research-backed questions/i)
     ).toBeInTheDocument();
   });
 
   it('has a take the quiz button', () => {
-    const { getByRole } = render(<ClimatePersonality />);
+    const { getByRole } = render(<StartQuiz />);
     expect(getByRole('button', { name: /take the quiz/i })).toBeInTheDocument();
   });
 });
