+++ src/pages/Home.tsx
@@ -4,6 +4,8 @@ import { Grid, Box, Typography, makeStyles } from '@material-ui/core';
 import { ReactComponent as Logo } from '../assets/cm-logo.svg';
 import ROUTES from '../components/Router/RouteConfig';
 import Wrapper from '../components/Wrapper';
+import PageSection from '../components/PageSection';
+import PageTitle from '../components/PageTitle';
 import Button from '../components/Button';
 import ScrollToTopOnMount from '../components/ScrollToTopOnMount';
 
@@ -32,103 +34,94 @@ const Home: React.FC<{}> = () => {
     <Grid container className={classes.root}>
       <ScrollToTopOnMount />
       <Wrapper bgColor="#82EFC5" justify="center">
-        <Grid
-          item
-          sm={12}
-          lg={4}
-          container
-          direction="column"
-          justify="center"
-          alignItems="center"
-        >
-          <Grid item>
-            <Box mb={2} mt={5}>
-              <Typography
-                variant="h4"
-                align="center"
-              >
-                Powering climate conversations
-              </Typography>
-            </Box>
-          </Grid>
+        <PageSection>
+          <Grid
+            item
+            container
+            direction="column"
+            justify="center"
+            alignItems="center"
+          >
+            <Grid item>
+              <Box mb={2} mt={5}>
+                <Typography variant="h4" align="center">
+                  Powering climate conversations
+                </Typography>
+              </Box>
+            </Grid>
 
-          <Grid item className={classes.logo}>
-            <Box mb={4}>
-              <Logo data-testid="climate-mind-logo" />
-            </Box>
-          </Grid>
+            <Grid item className={classes.logo}>
+              <Box mb={4}>
+                <Logo data-testid="climate-mind-logo" />
+              </Box>
+            </Grid>
 
-          <Grid item>
-            <Box ml={1} mr={1} mb={1}>
-              <Typography align="center">
-                I'll help you uncover your personal values to understand what
-                you care about most. Then we'll explore how climate change is
-                affecting your core values and the attractive solutions
-                available to address it.
-              </Typography>
-            </Box>
+            <Grid item>
+              <Box ml={1} mr={1} mb={1}>
+                <Typography align="center">
+                  I'll help you uncover your personal values to understand what
+                  you care about most. Then we'll explore how climate change is
+                  affecting your core values and the attractive solutions
+                  available to address it.
+                </Typography>
+              </Box>
+            </Grid>
           </Grid>
-        </Grid>
+        </PageSection>
       </Wrapper>
 
       <Wrapper bgColor="#07373B" justify="center">
-        <Grid
-          item
-          sm={12}
-          lg={4}
-          container
-          direction="column"
-          justify="center"
-          alignItems="center"
-        >
-          <Grid item>
-            <Box mb={3}>
-              <Typography
-                variant="h4"
-                align="center"
-                className={classes.bottomText}
-              >
-                What is Climate Mind?
-              </Typography>
-            </Box>
-          </Grid>
+        <PageSection>
+          <Grid
+            item
+            container
+            direction="column"
+            justify="center"
+            alignItems="center"
+          >
+            <Grid item>
+              <Box mb={3}>
+                <Typography
+                  variant="h4"
+                  align="center"
+                  className={classes.bottomText}
+                >
+                  What is Climate Mind?
+                </Typography>
+              </Box>
+            </Grid>
 
-          <Grid item>
-            <Box ml={1} mr={1} mb={3}>
-              <Typography
-                align="center"
-                className={classes.bottomText}
-              >
-                Climate Mind is a web app that helps you explore how your values
-                and personal interests are affected by climate change.
-              </Typography>
-            </Box>
-          </Grid>
+            <Grid item>
+              <Box ml={1} mr={1} mb={3}>
+                <Typography align="center" className={classes.bottomText}>
+                  Climate Mind is a web app that helps you explore how your
+                  values and personal interests are affected by climate change.
+                </Typography>
+              </Box>
+            </Grid>
 
-          <Grid item>
-            <Box ml={1} mr={1} mb={3}>
-              <Typography
-                align="center"
-                className={classes.bottomText}
-              >
-                We hope to inspire you to take action with a range of attractive
-                solutions consistent with your values that you can get excited
-                about.
-              </Typography>
-            </Box>
-          </Grid>
+            <Grid item>
+              <Box ml={1} mr={1} mb={3}>
+                <Typography align="center" className={classes.bottomText}>
+                  We hope to inspire you to take action with a range of
+                  attractive solutions consistent with your values that you can
+                  get excited about.
+                </Typography>
+              </Box>
+            </Grid>
 
-          <Grid item>
-            <Button
-              variant="contained"
-              color="primary"
-              disableElevation
-              onClick={() => history.push(ROUTES.ROUTE_PERSONALITY)}
-            >
-              Get Started
-            </Button>
+            <Grid item>
+              <Button
+                variant="contained"
+                color="primary"
+                disableElevation
+                onClick={() => history.push(ROUTES.ROUTE_PERSONALITY)}
+              >
+                Get Started
+              </Button>
+            </Grid>
           </Grid>
-        </Grid>
+        </PageSection>
       </Wrapper>
     </Grid>
   );
